instrumentation-loongsuite/loongsuite-instrumentation-agentscope/tests/requirements.txt
# Test requirements for latest versions
agentscope>=1.0.0
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-vcr>=1.0.2
vcrpy>=5.1.0
pyyaml>=6.0
opentelemetry-api>=1.37
opentelemetry-sdk>=1.37
opentelemetry-instrumentation>=0.58b0
opentelemetry-semantic-conventions>=0.58b0
opentelemetry-util-genai
wrapt

